refactor: extract id type and rename files
