--- packages/sanity/src/_singletons/context/MediaLibraryIdContext.ts
@@ -1,7 +1,0 @@
-import {createContext} from 'sanity/_createContext'
-
-/** @internal */
-export const MediaLibraryIdContext = createContext<{
-  libraryId: string
-  organizationId: string
-} | null>('sanity/_singletons/context/media-library', null)
