--- packages/sanity/src/core/form/studio/assetSourceMediaLibrary/hooks/useMediaLibraryId.tsx
@@ -1,6 +1,0 @@
-import {useContext} from 'react'
-import {MediaLibraryIdContext} from 'sanity/_singletons'
-
-export function useMediaLibraryIds(): {libraryId: string; organizationId: string} | null {
-  return useContext(MediaLibraryIdContext)
-}
